CDEC-433: Sin bin some more slotting code
