+++ sinbin/cardano-sl-sinbin.cabal
@@ -32,6 +32,7 @@ library
 
                         Pos.Sinbin.Slotting
                         Pos.Sinbin.Slotting.Class
+                        Pos.Sinbin.Slotting.MemState
                         Pos.Sinbin.Slotting.Types
                         Pos.Sinbin.Slotting.Util
                         Pos.Sinbin.Util.JsonLog.Events
