+++ sinbin/src/Pos/Sinbin/Slotting.hs
@@ -3,5 +3,6 @@ module Pos.Sinbin.Slotting
        ) where
 
 import           Pos.Sinbin.Slotting.Class as X
+import           Pos.Sinbin.Slotting.MemState as X
 import           Pos.Sinbin.Slotting.Types as X
 import           Pos.Sinbin.Slotting.Util as X
